feat: don't call tenant api on CE, so code split
