--- app/client/src/ce/sagas/tenantSagas.tsx
@@ -1,31 +1,0 @@
-import TenantApi from "@appsmith/api/TenantApi";
-import {
-  ReduxActionErrorTypes,
-  ReduxActionTypes,
-} from "@appsmith/constants/ReduxActionConstants";
-import { ApiResponse } from "api/ApiResponses";
-import { call, put } from "redux-saga/effects";
-import { validateResponse } from "sagas/ErrorSagas";
-
-export function* fetchCurrentTenantConfigSaga() {
-  try {
-    const response: ApiResponse = yield call(
-      TenantApi.fetchCurrentTenantConfig,
-    );
-    const isValidResponse: boolean = yield validateResponse(response);
-
-    if (isValidResponse) {
-      yield put({
-        type: ReduxActionTypes.FETCH_CURRENT_TENANT_CONFIG_SUCCESS,
-        payload: response.data,
-      });
-    }
-  } catch (error) {
-    yield put({
-      type: ReduxActionErrorTypes.FETCH_CURRENT_TENANT_CONFIG_ERROR,
-      payload: {
-        error,
-      },
-    });
-  }
-}
