+++ app/client/src/ee/sagas/tenantSagas.tsx
@@ -1,7 +1,28 @@
-export * from "ce/sagas/tenantSagas";
-import { ReduxActionTypes } from "ce/constants/ReduxActionConstants";
-import { fetchCurrentTenantConfigSaga } from "ce/sagas/tenantSagas";
-import { all, takeLatest } from "redux-saga/effects";
+import {
+  ReduxActionErrorTypes,
+  ReduxActionTypes,
+} from "ce/constants/ReduxActionConstants";
+import { PERMISSION_TYPE } from "pages/Applications/permissionHelpers";
+import { all, put, takeLatest } from "redux-saga/effects";
+
+// On CE we don't expose tenant config so this shouldn't make any API calls and should just return necessary permissions for the user
+export function* fetchCurrentTenantConfigSaga() {
+  try {
+    yield put({
+      type: ReduxActionTypes.FETCH_CURRENT_TENANT_CONFIG_SUCCESS,
+      payload: {
+        userPermissions: [PERMISSION_TYPE.CREATE_WORKSPACE],
+      },
+    });
+  } catch (error) {
+    yield put({
+      type: ReduxActionErrorTypes.FETCH_CURRENT_TENANT_CONFIG_ERROR,
+      payload: {
+        error,
+      },
+    });
+  }
+}
 
 export default function* tenantSagas() {
   yield all([
